Move LineFieldBuilder to index package
