+++ kifi-backend/modules/search/app/com/keepit/search/index/DocUtil.scala
@@ -4,7 +4,6 @@ import com.keepit.common.time._
 import com.keepit.search.index.graph.collection.CollectionIdList
 import com.keepit.search.index.graph.URIList
 import com.keepit.search.index.graph.Util
-import com.keepit.search.line.LineField
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
 import org.apache.lucene.analysis.tokenattributes.PositionIncrementAttribute
 import org.apache.lucene.analysis.tokenattributes.PayloadAttribute
